Make GTK+ DND use device-aware API.
Now the device attached to the GdkDragContext is used to perform device grabs
and such. The only pending point is about keyboard navigation, the X11 code
uses passive grabs and XEvent modifications to have certain key combinations
redirected to the ipc widget, but XGE/XI2 provides allocated events, which are
later freed before GDK gets to translate the event, so modifying the XI2 event
isn't going to work... Active grabs are used at the moment.